AP_Math: New Layout to make the header file nicer Just minor cosmetic enhancements. The platform specific implementation is now split.
